[et] Restructure tasks and helpers
