+++ tools/expotools/src/promote-packages/helpers.ts
@@ -1,7 +1,4 @@
 import chalk from 'chalk';
-import inquirer from 'inquirer';
-import readline from 'readline';
-import stripAnsi from 'strip-ansi';
 
 import { Parcel } from './types';
 import * as Changelogs from '../Changelogs';
@@ -22,50 +19,6 @@ export async function createParcelAsync(pkg: Package): Promise<Parcel> {
     gitDir: new GitDirectory(pkg.path),
     state: {},
   };
-}
-
-/**
- * Prompts the user to select packages to promote or demote.
- */
-export async function promptForPackagesToPromoteAsync(parcels: Parcel[]): Promise<string[]> {
-  const maxLength = parcels.reduce((acc, { pkg }) => Math.max(acc, pkg.packageName.length), 0);
-  const choices = parcels.map(({ pkg, state }) => {
-    const action = state.isDemoting ? red.bold('demoted') : green.bold('promoted');
-
-    return {
-      name: `will be ${green(pkg.packageName.padEnd(maxLength))} ${action} ${formatVersionChange(
-        state.versionToReplace,
-        pkg.packageVersion
-      )}`,
-      value: pkg.packageName,
-      checked: !state.isDemoting,
-    };
-  });
-  const { selectedPackageNames } = await inquirer.prompt([
-    {
-      type: 'checkbox',
-      name: 'selectedPackageNames',
-      message: 'Which packages do you want to promote?\n',
-      choices: [
-        // Choices unchecked by default (these being demoted) should be on top.
-        // We could sort them, but JS sorting algorithm is unstable :/
-        ...choices.filter((choice) => !choice.checked),
-        ...choices.filter((choice) => choice.checked),
-      ],
-      pageSize: Math.max(15, (process.stdout.rows ?? 15) - 15),
-    },
-  ]);
-  // Inquirer shows all those selected choices by name and that looks so ugly due to line wrapping.
-  // If possible, we clear everything that has been printed after the prompt.
-  if (process.stdout.columns) {
-    const bufferLength = choices.reduce(
-      (acc, choice) => acc + stripAnsi(choice.name).length + 2,
-      0
-    );
-    readline.moveCursor(process.stdout, 0, -Math.ceil(bufferLength / process.stdout.columns));
-    readline.clearScreenDown(process.stdout);
-  }
-  return selectedPackageNames;
 }
 
 /**
